+++ migrations/20150422195551_bairro.js
@@ -1,14 +1,15 @@
 'use strict';
 
 exports.up = function(knex, Promise) {
-    return knex.schema.createTable('bairro', function(table) {
-        table.increments('id_bairro').primary();
-        table.string('nome').notNullable();
-        table.boolean('ativo').notNullable();
-        table.bigInteger('cidade_id').notNullable().references('id_cidade').inTable('cidade');
-    });
+  return knex.schema.createTable('bairro', function(table) {
+    table.increments('id_bairro').primary();
+    table.string('nome').notNullable();
+    table.boolean('ativo').notNullable();
+    table.integer('cidade_id').notNullable()
+      .references('id_cidade').inTable('cidade');
+  });
 };
 
 exports.down = function(knex, Promise) {
-	return knex.schema.dropTable('bairro');
+  return knex.schema.dropTable('bairro');
 };
