+++ migrations/20150422195607_conta.js
@@ -7,7 +7,7 @@ exports.up = function(knex) {
     table.timestamp('data_fechamento');
     table.decimal('saldo', 18, 2).notNullable().defaultTo(0);
     table.boolean('ativo').notNullable().defaultTo(true);
-    table.bigInteger('pessoa_id').notNullable()
+    table.integer('pessoa_id').notNullable()
       .references('id_pessoa').inTable('pessoa');
   });
 };
