+++ migrations/20150422195615_ativacao.js
@@ -9,10 +9,10 @@ exports.up = function(knex) {
     table.decimal('longitude', 14, 10);
     table.decimal('altitude', 18, 10);
     table.boolean('ativo').notNullable();
-    table.bigInteger('usuario_pessoa_id').notNullable()
-        .references('pessoa_id').inTable('usuario');
-    table.bigInteger('veiculo_id').notNullable()
-    .references('id_veiculo').inTable('veiculo');
+    table.integer('usuario_pessoa_id').notNullable()
+      .references('pessoa_id').inTable('usuario');
+    table.integer('veiculo_id').notNullable()
+      .references('id_veiculo').inTable('veiculo');
   });
 };
