+++ migrations/20150422195619_usuario_credenciado.js
@@ -1,17 +1,17 @@
 'use strict';
 
 exports.up = function(knex, Promise) {
-    return knex.schema.createTable('usuario_credenciado', function(table) {
-        table.bigInteger('pessoa_fisica_pessoa_id')
-         	.primary()
-         	.references('pessoa_id').inTable('pessoa_fisica');
-        table.string('login').unique().notNullable();
-        table.string('senha');
-        table.boolean('primeiro_acesso').notNullable();
-        table.boolean('ativo').notNullable();
-    });
+  return knex.schema.createTable('usuario_credenciado', function(table) {
+    table.integer('pessoa_fisica_pessoa_id')
+      .primary()
+      .references('pessoa_id').inTable('pessoa_fisica');
+    table.string('login').unique().notNullable();
+    table.string('senha');
+    table.boolean('primeiro_acesso').notNullable();
+    table.boolean('ativo').notNullable();
+  });
 };
 
 exports.down = function(knex, Promise) {
-	return knex.schema.dropTable('usuario_credenciado');
+  return knex.schema.dropTable('usuario_credenciado');
 };
