+++ migrations/20150511122507_recuperacao_senha.js
@@ -1,13 +1,14 @@
 'use strict';
 
 exports.up = function(knex, Promise) {
-      return knex.schema.createTable('recuperacao_senha', function(table) {
-        table.uuid('id_recuperacao_senha').notNullable();
-        table.timestamp('data_expiracao').notNullable();
-        table.bigInteger('pessoa_id').notNullable().references('id_pessoa').inTable('pessoa');
-    });
+  return knex.schema.createTable('recuperacao_senha', function(table) {
+    table.uuid('id_recuperacao_senha').notNullable();
+    table.timestamp('data_expiracao').notNullable();
+    table.integer('pessoa_id').notNullable()
+      .references('id_pessoa').inTable('pessoa');
+  });
 };
 
 exports.down = function(knex, Promise) {
-	return knex.schema.dropTable('recuperacao_senha');
+  return knex.schema.dropTable('recuperacao_senha');
 };
